refactor(spec): split color/fill palette scale helpers by family
scale-color-helpers mixed generic ColorBrewer authoring with viridis,
gradient/steps, and hue/grey families. Extract those palette modules,
share gradient/steps stop resolvers, and keep package exports on the
scale-helpers facade. Knip alias allowlist extended for the new files.
