+++ knip.jsonc
@@ -15,6 +15,9 @@
     "packages/spec/src/scale-helpers.ts": ["duplicates"],
     "packages/spec/src/scale-position-helpers.ts": ["duplicates"],
     "packages/spec/src/scale-color-helpers.ts": ["duplicates"],
+    "packages/spec/src/scale-color-viridis.ts": ["duplicates"],
+    "packages/spec/src/scale-color-stops.ts": ["duplicates"],
+    "packages/spec/src/scale-color-hue-grey.ts": ["duplicates"],
     "packages/spec/src/scale-style-helpers.ts": ["duplicates"],
     "packages/spec/src/scale-multi-helpers.ts": ["duplicates"],
     "packages/spec/src/coord-helpers.ts": ["duplicates"],
