+++ packages/spec/src/scale-helpers.ts
@@ -3,7 +3,8 @@
  *
  * Implementation:
  *  - scale-position-helpers.ts — temporal, discrete, continuous, binned, reverse
- *  - scale-color-helpers.ts — color/fill families + configuredColorScaleType
+ *  - scale-color-helpers.ts — generic color/fill + ColorBrewer + configuredColorScaleType
+ *  - scale-color-viridis.ts / scale-color-stops.ts / scale-color-hue-grey.ts — palette families
  *  - scale-style-helpers.ts — numeric and finite-symbol style families
  *  - scale-multi-helpers.ts — multi-aes identity/manual + scaleType (#833)
  *
@@ -71,16 +72,6 @@ export {
   scaleColorBrewer,
   scaleColorDistiller,
   scaleColorFermenter,
-  scaleColorSteps,
-  scaleColorSteps2,
-  scaleColorStepsn,
-  scaleColorGradient,
-  scaleColorGradient2,
-  scaleColorGradientn,
-  scaleColorHue,
-  scaleColorGrey,
-  scaleColorGray,
-  scaleColorOrdinal,
   scaleColourContinuous,
   scaleColourDiscrete,
   scaleColourBinned,
@@ -93,16 +84,6 @@ export {
   scaleColourBrewer,
   scaleColourDistiller,
   scaleColourFermenter,
-  scaleColourSteps,
-  scaleColourSteps2,
-  scaleColourStepsn,
-  scaleColourGradient,
-  scaleColourGradient2,
-  scaleColourGradientn,
-  scaleColourHue,
-  scaleColourGrey,
-  scaleColourGray,
-  scaleColourOrdinal,
   scale_color_continuous,
   scale_color_discrete,
   scale_color_binned,
@@ -115,16 +96,6 @@ export {
   scale_color_brewer,
   scale_color_distiller,
   scale_color_fermenter,
-  scale_color_steps,
-  scale_color_steps2,
-  scale_color_stepsn,
-  scale_color_gradient,
-  scale_color_gradient2,
-  scale_color_gradientn,
-  scale_color_hue,
-  scale_color_grey,
-  scale_color_gray,
-  scale_color_ordinal,
   scale_colour_continuous,
   scale_colour_discrete,
   scale_colour_binned,
@@ -137,16 +108,6 @@ export {
   scale_colour_brewer,
   scale_colour_distiller,
   scale_colour_fermenter,
-  scale_colour_steps,
-  scale_colour_steps2,
-  scale_colour_stepsn,
-  scale_colour_gradient,
-  scale_colour_gradient2,
-  scale_colour_gradientn,
-  scale_colour_hue,
-  scale_colour_grey,
-  scale_colour_gray,
-  scale_colour_ordinal,
   scaleFillContinuous,
   scaleFillDiscrete,
   scaleFillBinned,
@@ -159,16 +120,6 @@ export {
   scaleFillBrewer,
   scaleFillDistiller,
   scaleFillFermenter,
-  scaleFillSteps,
-  scaleFillSteps2,
-  scaleFillStepsn,
-  scaleFillGradient,
-  scaleFillGradient2,
-  scaleFillGradientn,
-  scaleFillHue,
-  scaleFillGrey,
-  scaleFillGray,
-  scaleFillOrdinal,
   scale_fill_continuous,
   scale_fill_discrete,
   scale_fill_binned,
@@ -181,9 +132,20 @@ export {
   scale_fill_brewer,
   scale_fill_distiller,
   scale_fill_fermenter,
-  scale_fill_steps,
-  scale_fill_steps2,
-  scale_fill_stepsn,
+  type ColorScaleOptions,
+  type SequentialColorScaleOptions,
+  type BinnedColorScaleOptions,
+  type DiscreteColorScaleOptions,
+  type TransformedColorScaleOptions,
+  type TemporalColorScaleOptions,
+  type ManualColorScaleOptions,
+  type IdentityColorScaleOptions,
+  type ColorBrewerScaleOptions,
+  type ColorDistillerScaleOptions,
+  type ColorFermenterScaleOptions,
+} from "./scale-color-helpers.js";
+
+export {
   scaleColorViridisC,
   scaleColorViridisD,
   scaleColorViridisB,
@@ -202,36 +164,84 @@ export {
   scale_fill_viridis_c,
   scale_fill_viridis_d,
   scale_fill_viridis_b,
+  type ViridisScaleOptions,
+  type ViridisOptionName,
+} from "./scale-color-viridis.js";
+
+export {
+  scaleColorSteps,
+  scaleColorSteps2,
+  scaleColorStepsn,
+  scaleColorGradient,
+  scaleColorGradient2,
+  scaleColorGradientn,
+  scaleColourSteps,
+  scaleColourSteps2,
+  scaleColourStepsn,
+  scaleColourGradient,
+  scaleColourGradient2,
+  scaleColourGradientn,
+  scaleFillSteps,
+  scaleFillSteps2,
+  scaleFillStepsn,
+  scaleFillGradient,
+  scaleFillGradient2,
+  scaleFillGradientn,
+  scale_color_steps,
+  scale_color_steps2,
+  scale_color_stepsn,
+  scale_color_gradient,
+  scale_color_gradient2,
+  scale_color_gradientn,
+  scale_colour_steps,
+  scale_colour_steps2,
+  scale_colour_stepsn,
+  scale_colour_gradient,
+  scale_colour_gradient2,
+  scale_colour_gradientn,
+  scale_fill_steps,
+  scale_fill_steps2,
+  scale_fill_stepsn,
   scale_fill_gradient,
   scale_fill_gradient2,
   scale_fill_gradientn,
+  type StepsScaleOptions,
+  type Steps2ScaleOptions,
+  type StepsnScaleOptions,
+  type GradientScaleOptions,
+  type Gradient2ScaleOptions,
+  type GradientnScaleOptions,
+} from "./scale-color-stops.js";
+
+export {
+  scaleColorHue,
+  scaleColorGrey,
+  scaleColorGray,
+  scaleColorOrdinal,
+  scaleColourHue,
+  scaleColourGrey,
+  scaleColourGray,
+  scaleColourOrdinal,
+  scaleFillHue,
+  scaleFillGrey,
+  scaleFillGray,
+  scaleFillOrdinal,
+  scale_color_hue,
+  scale_color_grey,
+  scale_color_gray,
+  scale_color_ordinal,
+  scale_colour_hue,
+  scale_colour_grey,
+  scale_colour_gray,
+  scale_colour_ordinal,
   scale_fill_hue,
   scale_fill_grey,
   scale_fill_gray,
   scale_fill_ordinal,
-  type ColorScaleOptions,
-  type SequentialColorScaleOptions,
-  type BinnedColorScaleOptions,
-  type DiscreteColorScaleOptions,
-  type TransformedColorScaleOptions,
-  type TemporalColorScaleOptions,
-  type ManualColorScaleOptions,
-  type IdentityColorScaleOptions,
-  type ColorBrewerScaleOptions,
-  type ColorDistillerScaleOptions,
-  type ColorFermenterScaleOptions,
-  type StepsScaleOptions,
-  type Steps2ScaleOptions,
-  type StepsnScaleOptions,
-  type ViridisScaleOptions,
-  type ViridisOptionName,
-  type GradientScaleOptions,
-  type Gradient2ScaleOptions,
-  type GradientnScaleOptions,
   type HueScaleOptions,
   type GreyScaleOptions,
   type OrdinalColorScaleOptions,
-} from "./scale-color-helpers.js";
+} from "./scale-color-hue-grey.js";
 
 export * from "./scale-style-helpers.js";
